requirements.txt
praw
python-dotenv
pymongo[srv]
tkinter